added Widget Serializer. Added interface for CallbackKeeper
