+++ build.gradle
@@ -4,10 +4,10 @@ apply plugin: 'java'
 apply plugin: 'maven-publish'
 apply plugin: 'maven'
 apply plugin: 'idea'
-idea { module { downloadJavadoc = true; downloadSources = true; } }
+idea { module { downloadJavadoc = true; downloadSources = true } }
 
-Properties up = new Properties(); up.load(new FileInputStream("./version.properties"));
-project.version = up["liquidengine.legui.version"];
+Properties up = new Properties(); up.load(new FileInputStream("./version.properties"))
+project.version = up["liquidengine.legui.version"]
 
 configurations.all {
     resolutionStrategy.cacheChangingModulesFor 10, 'seconds'
@@ -88,40 +88,40 @@ project.sourceCompatibility = 1.8
 project.targetCompatibility = 1.8
 compileJava.options.encoding = 'UTF-8'
 
-def build_type = System.getProperty("build_type") ?: "";
+def build_type = System.getProperty("build_type") ?: ""
 if (build_type == "SNAPSHOT") {
     project.version += "-SNAPSHOT"
 }
 
 task publishArchives << {
-    def publishUrl;
-    def buildType;
-    def publishUsername = System.getenv("publishUsername") ?: publishUsername;
-    def publishPassword = System.getenv("publishPassword") ?: publishPassword;
-    def releasesPublishUrl = System.getenv("releasesPublishUrl") ?: releasesPublishUrl;
-    def snapshotsPublishUrl = System.getenv("snapshotsPublishUrl") ?: snapshotsPublishUrl;
+    def publishUrl
+    def buildType
+    def publishUsername = System.getenv("publishUsername") ?: publishUsername
+    def publishPassword = System.getenv("publishPassword") ?: publishPassword
+    def releasesPublishUrl = System.getenv("releasesPublishUrl") ?: releasesPublishUrl
+    def snapshotsPublishUrl = System.getenv("snapshotsPublishUrl") ?: snapshotsPublishUrl
     println("-------------------------------------")
     println("PUBLISH ARCHIVES TASK " + build_type)
     println("-------------------------------------")
     if (build_type != "") {
         boolean b = false; buildType = "ERROR"
         if (build_type == "RELEASE") {
-            publishUrl = releasesPublishUrl; buildType = "RELEASE"; b = true;
+            publishUrl = releasesPublishUrl; buildType = "RELEASE"; b = true
         }
         if (build_type == "SNAPSHOT") {
-            publishUrl = snapshotsPublishUrl; buildType = "SNAPSHOT"; b = true;
+            publishUrl = snapshotsPublishUrl; buildType = "SNAPSHOT"; b = true
         }
         println("build_type : $buildType; PUBLISHED TO $publishUrl")
         if (b) {
             uploadArchives {
                 repositories.mavenDeployer {
-                    configuration = configurations.deployerJar;
+                    configuration = configurations.deployerJar
                     repository(url: publishUrl) {
                         authentication(userName: publishUsername, password: publishPassword)
                     }
                 }
             }
-            uploadArchives.execute();
+            uploadArchives.execute()
         }
     } else {
         println("NOT UPLOADED")
