+++ src/main/java/org/liquidengine/legui/serialize/json/gson/GsonSerializeUtil.java
@@ -37,13 +37,13 @@ public final class GsonSerializeUtil {
         return serializer.jsonSerialize(component, context);
     }
 
-    public static Object deserializeFromJson(JsonObject json) {
+    public static <T> T deserializeFromJson(JsonObject json) {
         return deserializeFromJson(json, new GsonSerializeContext());
     }
 
-    public static Object deserializeFromJson(JsonObject json, GsonSerializeContext context) {
+    public static <T> T deserializeFromJson(JsonObject json, GsonSerializeContext context) {
         String type = getType(json);
-        AbstractGsonSerializer serializer = GsonSerializeRegistry.getRegistry().getSerializer(type);
+        AbstractGsonSerializer<T> serializer = GsonSerializeRegistry.getRegistry().getSerializer(type);
         if (serializer == null) throw new LeguiException(LeguiExceptions.DESERIALIZER_IS_NOT_EXIST.message(type));
         return serializer.jsonDeserialize(json, context);
     }
