+++ src/main/java/org/liquidengine/legui/serialize/json/gson/component/GsonButtonSerializer.java
@@ -30,7 +30,7 @@ public class GsonButtonSerializer extends GsonComponentSerializer<Button> {
         JsonElement textState = json.get(TEXT_STATE);
         if (isNotNull(textState)) {
             JsonObject asJsonObject = textState.getAsJsonObject();
-            TextState state = (TextState) GsonSerializeUtil.deserializeFromJson(asJsonObject, context);
+            TextState state = GsonSerializeUtil.deserializeFromJson(asJsonObject, context);
             object.getTextState().copy(state);
         }
     }
