+++ src/main/java/org/liquidengine/legui/serialize/json/gson/component/optional/GsonTextStateSerializer.java
@@ -6,11 +6,10 @@ import org.liquidengine.legui.component.optional.TextState;
 import org.liquidengine.legui.component.optional.align.HorizontalAlign;
 import org.liquidengine.legui.component.optional.align.VerticalAlign;
 import org.liquidengine.legui.serialize.json.gson.AbstractGsonSerializer;
-import org.liquidengine.legui.serialize.json.gson.GsonUtil;
 import org.liquidengine.legui.serialize.json.gson.GsonSerializeContext;
+import org.liquidengine.legui.serialize.json.gson.GsonUtil;
 
 import static org.liquidengine.legui.serialize.json.gson.GsonConstants.*;
-import static org.liquidengine.legui.serialize.json.gson.GsonConstants.FONT;
 import static org.liquidengine.legui.serialize.json.gson.GsonUtil.isNotNull;
 import static org.liquidengine.legui.serialize.json.gson.GsonUtil.readColor;
